challenges/snek/expl_requirements.txt
pwntools
pillow
